+++ test/jdk/jdk/incubator/code/lib/resolve/Parent.java
@@ -20,7 +20,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package lib;
+package resolve;
 
 public class Parent {
